[ui] Clean up PipelineRoot.test
