+++ js_modules/dagit/packages/core/src/launchpad/LaunchpadTransientSessionContainer.tsx
@@ -9,12 +9,12 @@ import {
 import {useFeatureFlags} from '../app/Flags';
 import {RepoAddress} from '../workspace/types';
 
-import {LaunchpadType} from './LaunchpadRoot';
 import LaunchpadSession from './LaunchpadSession';
+import {LaunchpadType} from './types';
 import {
   LaunchpadSessionPartitionSetsFragment,
   LaunchpadSessionPipelineFragment,
-} from './types/LaunchpadRoot.types';
+} from './types/LaunchpadAllowedRoot.types';
 
 interface Props {
   launchpadType: LaunchpadType;
