+++ js_modules/dagit/packages/core/src/launchpad/types.tsx
@@ -1,0 +1,1 @@
+export type LaunchpadType = 'asset' | 'job';
