+++ js_modules/dagit/packages/core/src/pipelines/__tests__/PipelineRoot.test.tsx
@@ -1,10 +1,28 @@
-import {act, render, screen, waitFor} from '@testing-library/react';
+import {render, screen} from '@testing-library/react';
 import * as React from 'react';
+import {MemoryRouter} from 'react-router-dom';
 
-import {TestProvider} from '../../testing/TestProvider';
+import {TestPermissionsProvider} from '../../testing/TestPermissions';
 import {buildRepoAddress} from '../../workspace/buildRepoAddress';
 import {repoAddressAsURLString} from '../../workspace/repoAddressAsString';
 import {PipelineRoot} from '../PipelineRoot';
+
+jest.mock('../../launchpad/LaunchpadAllowedRoot', () => ({
+  LaunchpadAllowedRoot: () => <div>launchpad allowed placeholder</div>,
+}));
+
+jest.mock('../PipelineOverviewRoot', () => ({
+  PipelineOverviewRoot: () => <div>pipeline overview placeholder</div>,
+}));
+
+jest.mock('../PipelineOrJobDisambiguationRoot', () => ({
+  PipelineOrJobDisambiguationRoot: () => <div>pipeline or job disambiguation placeholder</div>,
+}));
+
+jest.mock('../../nav/PipelineNav', () => ({
+  ...jest.requireActual('../../nav/PipelineNav'),
+  PipelineNav: () => <div />,
+}));
 
 jest.mock('../GraphExplorer', () => ({
   ...jest.requireActual('../GraphExplorer'),
@@ -13,47 +31,18 @@ jest.mock('../GraphExplorer', () => ({
   GraphExplorer: () => <div />,
 }));
 
+jest.mock('../../app/analytics', () => ({
+  ...jest.requireActual('../../app/analytics'),
+  useTrackPageView: jest.fn(),
+}));
+
 // This file must be mocked because Jest can't handle `import.meta.url`.
 jest.mock('../../graph/asyncGraphLayout', () => ({}));
 
 const REPO_NAME = 'foo';
 const REPO_LOCATION = 'bar';
-const PIPELINE_NAME = 'pipez';
 
 describe('PipelineRoot', () => {
-  const mocks = {
-    Pipeline: () => ({
-      id: () => PIPELINE_NAME,
-      modes: () => [...new Array(1)],
-      isAssetJob: () => false,
-    }),
-    SolidDefinition: () => ({
-      assetNodes: () => [],
-    }),
-    PipelineSnapshot: () => ({
-      runs: () => [],
-      schedules: () => [],
-      sensors: () => [],
-      solidHandle: null,
-    }),
-    RepositoryLocation: () => ({
-      id: REPO_LOCATION,
-      name: REPO_LOCATION,
-      repositories: () => [...new Array(1)],
-    }),
-    Repository: () => ({
-      id: REPO_NAME,
-      name: REPO_NAME,
-      pipelines: () => [...new Array(1)],
-      assetGroups: () => [...new Array(0)],
-    }),
-    DagitQuery: () => ({
-      assetNodes: () => [], // standard job, not an asset job
-    }),
-  };
-
-  const apolloProps = {mocks};
-
   const repoAddress = buildRepoAddress(REPO_NAME, REPO_LOCATION);
   const pipelineName = 'pipez';
   const path = `/locations/${repoAddressAsURLString(
@@ -61,75 +50,51 @@ describe('PipelineRoot', () => {
   )}/pipelines/${pipelineName}:default`;
 
   it('renders overview by default', async () => {
-    const routerProps = {
-      initialEntries: [path],
+    render(
+      <MemoryRouter initialEntries={[path]}>
+        <PipelineRoot repoAddress={repoAddress} />
+      </MemoryRouter>,
+    );
+
+    const overviewDummy = await screen.findByText(/pipeline overview placeholder/i);
+    expect(overviewDummy).toBeVisible();
+  });
+
+  it('renders playground route', async () => {
+    const locationPermissions = {
+      [REPO_LOCATION]: {
+        canLaunchPipelineExecution: {enabled: true, disabledReason: ''},
+      },
     };
 
     render(
-      <TestProvider apolloProps={apolloProps} routerProps={routerProps}>
-        <PipelineRoot repoAddress={repoAddress} />
-      </TestProvider>,
+      <TestPermissionsProvider locationOverrides={locationPermissions}>
+        <MemoryRouter initialEntries={[`${path}/playground`]}>
+          <PipelineRoot repoAddress={repoAddress} />
+        </MemoryRouter>
+      </TestPermissionsProvider>,
     );
 
-    await waitFor(() => {
-      const selected = screen.getByRole('tab', {selected: true});
-      expect(selected.textContent).toMatch(/overview/i);
-    });
+    const playgroundDummy = await screen.findByText(/launchpad allowed placeholder/i);
+    expect(playgroundDummy).toBeVisible();
   });
 
-  describe('Limits launchpad based on launch permission', () => {
-    it('renders launchpad route and tab by default', async () => {
-      const routerProps = {
-        initialEntries: [`${path}/playground`],
-      };
+  it('redirects to disambiguation if no launch permission', async () => {
+    const locationPermissions = {
+      [REPO_LOCATION]: {
+        canLaunchPipelineExecution: {enabled: false, disabledReason: 'no can do'},
+      },
+    };
 
-      act(() => {
-        render(
-          <TestProvider apolloProps={apolloProps} routerProps={routerProps}>
-            <PipelineRoot repoAddress={repoAddress} />
-          </TestProvider>,
-        );
-      });
-
-      const selected = await screen.findByRole('tab', {selected: true});
-
-      // Route to Playground, verify that the "New run" tab appears.
-      expect(selected.textContent).toMatch(/launchpad/i);
-
-      await waitFor(() => {
-        expect(screen.getByText(/\+ add/i)).toBeVisible();
-      });
-    });
-
-    it('redirects playground route if no launch permission', async () => {
-      const routerProps = {
-        initialEntries: [`${path}/playground`],
-      };
-
-      render(
-        <TestProvider
-          permissionOverrides={{
-            launch_pipeline_execution: {enabled: false, disabledReason: 'nope'},
-          }}
-          apolloProps={apolloProps}
-          routerProps={routerProps}
-        >
+    render(
+      <TestPermissionsProvider locationOverrides={locationPermissions}>
+        <MemoryRouter initialEntries={[`${path}/playground`]}>
           <PipelineRoot repoAddress={repoAddress} />
-        </TestProvider>,
-      );
+        </MemoryRouter>
+      </TestPermissionsProvider>,
+    );
 
-      const selected = await screen.findByRole('tab', {selected: true});
-
-      // Redirect to Definition, which has been highlighted in the tabs.
-      expect(selected.textContent).toMatch(/overview/i);
-
-      // Render a disabled "Launchpad" tab.
-      await waitFor(() => {
-        expect(screen.queryByRole('tab', {name: /launchpad/i})).toHaveAttribute(
-          'aria-disabled',
-          'true',
-        );
-      });
-    });
+    const overviewDummy = await screen.findByText(/pipeline or job disambiguation placeholder/i);
+    expect(overviewDummy).toBeVisible();
   });
 });
